working, but with concrete T.caller = Int64 and faked API call

+++ src/as_types/type.ml
@@ -107,7 +107,7 @@ let catch = Prim Error
 
 (* Shared call context *)
 
-let caller = Prim Int32 (*TBR*)
+let caller = Prim Int64 (*TBR*)
 let ctxt = Obj (Object,[{ lab = "caller"; typ = caller }])
 
 let prim = function
